requirements.txt
Django==6.0
channels==4.3.2
daphne==4.2.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
Pillow==12.1.1
django-filter==25.2
whitenoise==6.9.0
